Add entity disambiguation utilities and enhance canonical entity processing
- Introduced new utility functions for entity disambiguation in `fuzzy.py`.
- Implemented `assign_label_instances` and `map_canonical_entities_ner_preds` in `core.py`.
- Added LLM inference capabilities in `llm.py` for refining canonical entities.
- Updated `entities.py` to include `aymurai_label_instance` for ordered label indexing.
