+++ aymurai/utils/entity_disambiguation/__init__.py
@@ -1,0 +1,19 @@
+from aymurai.utils.entity_disambiguation.core import (
+    assign_label_instances,
+    map_canonical_entities_ner_preds,
+)
+from aymurai.utils.entity_disambiguation.fuzzy import (
+    build_canonical_entities,
+)
+from aymurai.utils.entity_disambiguation.llm import (
+    llm_canonical_entities_inference,
+    load_prompts_from_yaml,
+)
+
+__all__ = [
+    "assign_label_instances",
+    "build_canonical_entities",
+    "llm_canonical_entities_inference",
+    "map_canonical_entities_ner_preds",
+    "load_prompts_from_yaml",
+]
